requirements.txt
pytest==3.6.4
webdriver-manager==1.7
selenium==3.14.0
allure-pytest==2.5.0
pytest-xdist==1.22.5